+++ src/App.tsx
@@ -7,10 +7,9 @@ import './assets/styles/App.scss'
 import { store } from './app/store'
 import * as ROUTES from './constants/routes'
 
-// import { UserProvider } from './providers/UserProvider'
 import { Header } from './components/Header/Header'
 import { Homepage } from './pages/Homepage/Homepage'
-import { ForgotPassword } from './pages/ForgotPassword'
+import { ForgotPassword } from './pages/ForgotPassword/ForgotPassword'
 import { LoginPage } from './pages/Login/LoginPage'
 import { SignUpPage } from './pages/Signup/Signup'
 import { Track } from './pages/Track/Track'
